// src/app/components/klpt-learning-observation-tool/models/klpt-domain.ts
import { KlptElement } from './klpt-element';
import { KlptSubDomain } from './klpt-sub-domain';

export interface KlptReflectiveQuestion {
  id: string;
  index: number;
  question: string;
}

export interface KlptDomain {
  id: string;
  index: number;
  name: string;
  summary: string;
  subDomains?: KlptSubDomain[];
  elements?: KlptElement[];
  reflectiveQuestions?: KlptReflectiveQuestion[];
}